Make trie.Storage manage the root key
Previously, this was done by the caller of `trie.Root`, leading to
duplicated code.

Note that `trie.Root` now requires write access to `trie.Storage`
so it can update the root key. However, since we plan to make
`state.Root` lazy like `trie.Root` (and thus require write access
when the state commitment needs to be recalculated), it shouldn't
matter.